Change icon to generic AWS icon, fix right click on table popup
